Remove the GrpcMethod object that is not really necessary at the moment
